Namespace JRoute class
