+++ libraries/import.legacy.php
@@ -55,7 +55,6 @@ JLoader::registerPrefix('J', JPATH_PLATFORM . '/legacy');
 
 // Register classes that don't follow one file per class naming conventions.
 JLoader::register('JText', JPATH_PLATFORM . '/joomla/language/text.php');
-JLoader::register('JRoute', JPATH_PLATFORM . '/joomla/application/route.php');
 
 // Check if the JsonSerializable interface exists already
 if (!interface_exists('JsonSerializable'))
