+++ libraries/src/Joomla/CMS/Router/Router.php
@@ -12,6 +12,7 @@ defined('JPATH_PLATFORM') or die;
 
 use Joomla\CMS\Application\ApplicationHelper;
 use Joomla\CMS\Component\ComponentHelper;
+use Joomla\CMS\Uri\Uri;
 
 /**
  * Class to create and parse routes
@@ -194,6 +195,69 @@ class Router
 		}
 
 		return self::$instances[$client];
+	}
+
+	/**
+	 * Translates an internal Joomla URL to a humanly readable URL.
+	 *
+	 * @param   string   $url    Absolute or Relative URI to Joomla resource.
+	 * @param   boolean  $xhtml  Replace & by &amp; for XML compliance.
+	 * @param   integer  $ssl    Secure state for the resolved URI.
+	 *                             0: (default) No change, use the protocol currently used in the request
+	 *                             1: Make URI secure using global secure site URI.
+	 *                             2: Make URI unsecure using the global unsecure site URI.
+	 *
+	 * @return string The translated humanly readable URL.
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public function _($url, $xhtml = true, $ssl = null)
+	{
+		if (!is_array($url) && (strpos($url, '&') !== 0) && (strpos($url, 'index.php') !== 0))
+		{
+			return $url;
+		}
+
+		// Build route.
+		$uri = $this->build($url);
+
+		$scheme = array('path', 'query', 'fragment');
+
+		/*
+		 * Get the secure/unsecure URLs.
+		 *
+		 * If the first 5 characters of the BASE are 'https', then we are on an ssl connection over
+		 * https and need to set our secure URL to the current request URL, if not, and the scheme is
+		 * 'http', then we need to do a quick string manipulation to switch schemes.
+		 */
+		if ((int) $ssl || $uri->isSsl())
+		{
+			static $host_port;
+
+			if (!is_array($host_port))
+			{
+				$uri2 = Uri::getInstance();
+				$host_port = array($uri2->getHost(), $uri2->getPort());
+			}
+
+			// Determine which scheme we want.
+			$uri->setScheme(((int) $ssl === 1 || $uri->isSsl()) ? 'https' : 'http');
+			$uri->setHost($host_port[0]);
+			$uri->setPort($host_port[1]);
+			$scheme = array_merge($scheme, array('host', 'port', 'scheme'));
+		}
+
+		$url = $uri->toString($scheme);
+
+		// Replace spaces.
+		$url = preg_replace('/\s/u', '%20', $url);
+
+		if ($xhtml)
+		{
+			$url = htmlspecialchars($url, ENT_COMPAT, 'UTF-8');
+		}
+
+		return $url;
 	}
 
 	/**
